added bg image to main menu
